day02-python-url-shortener/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
validators==0.20.0
pytest==7.3.2
requests==2.31.0
Flask-Migrate==4.0.4
Flask-Limiter==2.8.0
